Extract math functions into separate files
